proto: factor out MultifileErrorCollector
This class passed the "three strikes and you refactor" test.
Also I'm about to add a fourth one soon. Too many.
