+++ gn/BUILD.gn
@@ -229,9 +229,9 @@ protobuf_full_deps_allowlist = [
   "../buildtools/grpc:*",
   "../src/ipc/protoc_plugin:*",
   "../src/protozero/protoc_plugin:*",
+  "../src/protozero:multifile_error_collector",
   "../src/protozero/filtering:filter_util",
   "../src/trace_processor:trace_processor_shell_lib",
-  "../src/protozero/filtering:filter_util",
   "../tools/*",
   "../src/tools/*",
   "../contrib/rust-sdk/tools/*",
